Extended UriReport into two new types
We currently have the idea of UriReport for reporting on a
particular URI (duh). This makes great sense when coupled with
an overall PerformanceReport object (which contains multiple
UriReports). However, UriReports is made up of HttpSamples and
this only makes sense when samples are available.

Well you could just say, samples must always be available! But
this isn't currently the case. We currently support two summarized
formats (JMeterSummarizerParser & WrkSummarizerParser). So how do
we accomodate them? Well currently we add "summarized" fields to
HttpSample and deal with it that way. This is terrible not only
because a single HttpSample can represent summarized data, but
also becuase now everything has to be concerned with working directly
with HttpSamples which is a terribly coupled interface to expose.

So I've now made UriReport abstract and split it into two concrete
sub-classes AccumulatedUriReport and AggregateUriReport. The Accumulated
version contains the original HttpSamples while the Aggregate contains,
well, aggregate information.

The Summarizer parsers have now been updated to use this new format and
I believe that this results in easier to understand code-base. There is
still more work to do (hiding HttpSamples all together, etc.). But
this is pushing us in the correct direction.