Clean Up - Stop Accessing HttpSamples
Removed a few different places using UriReport.getHttpSamples
mainly because that is very wrong to have to reach so many
layers deep to get what you need. Especially now that we have
different URI reports.

I would like to make the field private, but the current
PerformanceReportTest.java is holding me up because it has some
bad usage and test-cases that need to be re-thought before changing.